@@ -43,44 +43,47 @@
     "iframe"
   ],
   "peerDependencies": {
-    "react": "^0.14.0 || >=15.0.0"
+    "react": ">=16.3.0"
   },
   "dependencies": {
     "prop-types": "^15.5.8"
   },
   "devDependencies": {
-    "@trendmicro/react-buttons": "~1.0.3",
-    "babel-cli": "~6.24.0",
-    "babel-core": "~6.24.0",
-    "babel-eslint": "~7.2.2",
-    "babel-loader": "~6.4.0",
-    "babel-plugin-transform-decorators-legacy": "~1.3.4",
-    "babel-preset-es2015": "~6.24.0",
-    "babel-preset-react": "~6.24.1",
-    "babel-preset-stage-0": "~6.24.1",
-    "coveralls": "~2.13.0",
-    "css-loader": "~0.28.0",
-    "eslint": "~4.5.0",
-    "eslint-config-trendmicro": "~1.0.0",
-    "eslint-loader": "~1.7.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "^7.3.0",
-    "file-loader": "~0.11.1",
-    "html-webpack-plugin": "~2.28.0",
-    "http-server": "~0.9.0",
+    "@babel/cli": "~7.1.2",
+    "@babel/core": "~7.1.2",
+    "@babel/polyfill": "~7.0.0",
+    "@babel/preset-env": "~7.1.0",
+    "@babel/preset-react": "~7.0.0",
+    "@babel/register": "~7.0.0",
+    "@trendmicro/babel-config": "~1.0.0-alpha",
+    "babel-eslint": "~10.0.1",
+    "babel-loader": "~8.0.4",
+    "coveralls": "~3.0.2",
+    "css-loader": "~1.0.1",
+    "eslint": "~5.10.0",
+    "eslint-config-trendmicro": "~1.4.1",
+    "eslint-loader": "~2.1.1",
+    "eslint-plugin-import": "~2.14.0",
+    "eslint-plugin-jsx-a11y": "~6.1.2",
+    "eslint-plugin-react": "~7.11.1",
+    "file-loader": "~3.0.1",
+    "html-webpack-plugin": "~3.2.0",
+    "http-server": "~0.11.1",
+    "mini-css-extract-plugin": "^0.5.0",
     "nib": "~1.1.2",
-    "react": "^0.14.0 || >=15.0.0",
-    "react-dom": "^0.14.0 || >=15.0.0",
-    "style-loader": "~0.16.1",
+    "react": ">=16.3.0",
+    "react-dom": ">=16.3.0",
+    "react-styleguidist": "~8.0.6",
+    "resize-observer-polyfill": "~1.5.1",
+    "style-loader": "~0.23.1",
     "stylint": "~1.5.9",
-    "stylint-loader": "~1.0.0",
-    "stylus-loader": "~3.0.1",
-    "tap": "~10.3.0",
-    "trendmicro-ui": "~0.4.0",
-    "url-loader": "~0.5.8",
-    "webpack": "~2.4.1",
-    "webpack-dev-server": "~2.4.1",
-    "which": "~1.2.12"
+    "stylus-loader": "~3.0.2",
+    "tap": "~12.1.1",
+    "trendmicro-ui": "~0.5.2",
+    "url-loader": "~1.1.2",
+    "webpack": "~4.28.1",
+    "webpack-cli": "~3.1.2",
+    "webpack-dev-server": "~3.1.10",
+    "which": "~1.3.1"
   }
 }
