@@ -1,3 +1,3 @@
 import Iframe from './Iframe';
 
-module.exports = Iframe;
+export default Iframe;
